[FIX] fic

+++ sofraco-back-end/src/services/utils/files.js
@@ -26,9 +26,11 @@ exports.deleteFilesinDirectory = (directory) => {
         if (err) throw err;
 
         for (const file of files) {
-            fs.unlink(path.join(directory, file), err => {
-                if (err) throw err;
-            });
+            if (file.includes('.')) {
+                fs.unlink(path.join(directory, file), err => {
+                    if (err) throw err;
+                });
+            }
         }
     });
 }
